requirements.txt
pillow
pytesseract
pydub
speechrecognition
transformers
